test: add tpch 3, 15, 17 planner test (#86)
After #80, we can test subqueries with DF logical optimizer enabled. The
reasons not all queries are working are:

- Multiple join filters, possibly containing non-eq predicate.
- optd panicking with `cannot find best binding for group`. I suspect
this is due to the partial explore?

In addition, I realized I pasted the wrong query for q18 🥵 (that made me
double-checked whether other queries are correct), so it is removed now.