feat: add RuleWrapper and make the optimizer a hybrid optimizer (#88)
# Design
This pr is for adding a rule wrapper and hybrid logic to make cascades
optimizer support two types of logic rules: rules registered as
heuristics and rules registered as cascades.
- heuristic rule, which should always replace the old expr using the new
expr if rule is applied
- heuristic rule should only return 1 or 0 expr, if not match, return
the 0 expr
- heuristic rule's newly generated expr would be an expr not explored
yet, only marks the heuristic rule which generates it as fired
- heuristic rule's newly generated expr might have been existed in memo,
in that case two groups should be merged and old expr should be marked
as a dead end (all rules fired for it)

# Major Changes
- Add RuleWrapper and optimizeType
- Change OptimizeExpression Task and ApplyRuleTask logic
- Add replace_group_expr for memo and optimizer

# why we are using a hybrid optimizer instead of two stage optimizers
including heuristics and cascades?
Currently the columnRef property define emptyRelation derive rules as
derive instead of a real column ref. However, this is not true for a
two-stage optimizer. For instance, if we have a query like this select *
from t1 join t2 on false, the original plan tree is Projection(columnRef
0,1,2,3) -> InnerJoin -> TableScan, after heuristic rules it becomes
Projection(columnRef 0,1,2,3) -> EmptyRelation. But the emptyRelation
has no connected table name or properties containing its schema(even if
we create infer properties rules in heuristics, CascadesOptimizer cannot
get the properties generated by heuristicOptimizers), but the projection
node require columnref 0,1,2,3, so the columnRef derive rule will panic
as it cannot infer from the children of emptyRelation. If we put
heuristic rules in cascades optimizer, it works fine as the original
table scan is stored in original plan and can be fetched from catalog,
after transformation, the same group share the same properties, so it
works fine.

If we're going to create a two-stage optimizer, it is hard to share
properties/schema across two optimizers. No matter we're modifying
current heuristic implementations with infer properties or creating a
heuristic optimizer from cascades framework with customized task.
(because the schema are either stored in memo or from catalog with table
name, it cannot be easily shared with outside calls)

# future tasks
the lack of maintaining properties in remembered state might break the
adaptive design as the emptyRelation is record and no connected schema
without inferred properties.

---------

Signed-off-by: AveryQi115 <averyqi115@gmail.com>