feat: load tpch into postgres (#95)
**Summary**: Code for installing/initializing/running Postgres as well
as loading TPC-H into Postgres.

**Demo**:

TPC-H tables and queries are already generated. In the first run, we
load the TPC-H data into Postgres. In the second run, we just use the
already loaded `pgdata/`.


https://github.com/cmu-db/optd/assets/20631215/a412185c-7c61-436f-9232-2b0d61863c43


**Details**:
* Postgres installation/initialization/running is all cached and only
done if necessary
* Loading benchmarks is cached as well and is based on the configuration
of the benchmark _only if the benchmark is read-only_
* As before, caching logic is atomic
* Benchmarks and benchmark configs are extensible, including the caching
logic (via `get_strid()`)