Feat: flatten filter and join cond to list of expr, simplify filter rule (#108)
# Major Changes
- BinOpExpr in Join cond and filter cond is changed to LogOpExpr
- LogOpExpr child is not a single ExprList but multiple exprs as
children (which is different from previous logOpExpr)
- ColumnRef and Schema properties changed for LogOpExpr

# Design
Previously, LogOpExpr is designed to be `children: vec![ExprList]` which
means the LogOpExpr only has one child of ExprList. And condition
expressions in filter and join from datafusion will be converted to
BinOpExpr, which means the LogOpExpr actually does not exist in optd.
After converting conditions from datafusion to be LogOpExpr, the design
of one child LogOpExpr will influence the ProjectPullUpJoinRule and
other rules with schema recalculated as these rules simply traverse the
expr's children in its relNode format, it will treat LogOpExpr children
as only one child (ExprList), and when it further traverses ExprList it
has no idea whether it is a LogOp list or a normal list.

So current design change the LogOpExpr to include multiple children to
represent its real children size in RelNode format.
It has further impact on cost model and properties builder logic as
previously List is the only type in Optd which contains multiple
children.

About the pickMany stuff, it won't change the core's logic no matter it has one or many children as we only uses as join cond or filter cond, their cond is only one expr which is LogOpExpr and LogOpExpr is expanded and returned with all of it children. So the LogOpExpr already play the role as a one child wrapper(like ExprList), no need to add a ExprList.

#TODO
- [x] currently I simply copy properties derive and build rules for List
as the rules for LogOpExpr, please check whether it is aligned with cost
model's logic
- [x] the change break the selectivity infer rules and cost model
calculation so it cannot find a best group bindings for several cases.
Please update the selectivity calculation for logOpExpr.

---------

Signed-off-by: AveryQi115 <averyqi115@gmail.com>
Co-authored-by: Patrick Wang <wang.patrick57@gmail.com>