Fix: fix for heuristic rule wrapper (#147)
The new expr returned by heuristic rules are not in the original group,
which means it never got an OptimizeExpressionTask with exploring as
false (OptimizeExpressionTask with exploring=False only be called in
OptimizeGroup), it should evoke an OptimizeExpressionTask with
exploring=false for itself to apply all the transform rules and
implementation rules for itself.

Besides, the pr adds checks for whether the original expr equals to the
output expr for heuristic rule. If that's the case, it should prompt an
error as this breaks the heuristic rule's definition. (Silently
accepting it will mark the input expr being as a dead end and there's no
more new exprs to replace it)

Signed-off-by: AveryQi115 <averyqi115@gmail.com>