fix: reference the correct column stats for nlj and hash join (#164)
The indices of `ColumnRefExpr`s in hash join columns are with respect to
left or right tables. For example,


![image](https://github.com/cmu-db/optd/assets/52783948/b2ea40e1-695b-4406-8231-0d7a4020bb63)

The plan in this figure means join on 0th column of the left table and
3rd column of the right table.

But for NLJ, the column indices are with respect to the join output. For
instance, if the condition is `#0 = #5` and the left table has 3
columns, the condition actually means join on 0th column of the left
table and 3rd column of the right.