fix(tests): avoid using df logical optimizer as possible (#240)
There was a time that the optd optimizer itself does not handle a lot of
queries and we have to rely on the datafusion logical optimizer. Now we
have our own logical optimizer and we can drop `use_df_logical` in a lot
of places. The exceptions are TPC-H queries, where we still need to
finish the subquery unnesting work before fully opt-in optd logical
optimizer.

Signed-off-by: Alex Chi <iskyzh@gmail.com>